Refactor audio effects module with improved JSI installation and imports
