+++ shared/Audio/effects/components/Delay.hpp
@@ -10,7 +10,7 @@
 #include <type_traits>
 
 #include "EffectBase.hpp"
-#include "EffectConstants.hpp"
+#include "constant/EffectConstants.hpp"
 
 namespace AudioFX {
