+++ shared/Audio/effects/managers/DelayManager.h
@@ -2,10 +2,12 @@
 
 #include <atomic>
 #include <memory>
+#include <mutex>
 #include <string>
 #include <vector>
 
-#include "../../../jsi/JSICallbackManager.h"
+#include "../../capture/jsi/JSICallbackManager.h"
+#include "../components/Delay.hpp"
 #include "../config/EffectsConfig.h"
